+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project
 adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] - 2020-01-02
+## [Unreleased] - 2020-01-12
 ### Added
 - A Dockerfile was added at
   [resources/docker/rust-linux-x86_64/Dockerfile](resources/docker/rust-linux-x86_64/Dockerfile)
@@ -27,6 +27,8 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
   of initializing the FFI code for plugins so that plugin authors do not have to.
 - Callbacks were added to the `kpal-plugin` library, making it easier to isolate the parts of the
   plugin code that communicate with the hardware.
+- Peripherals may how have attributes that contain any character within the ASCII character set
+  except for the null byte.
 
 ### Changed
 - All artifacts are now built on the CI with the `--release` profile.